refactor: rename static folder to constnts
