+++ src/js/Keyboard/constants/KeyCase.js
