+++ src/js/Keyboard/constants/KeyType.js
