+++ src/js/Keyboard/constants/KeyboardSettings.js
@@ -1,0 +1,4 @@
+import { language } from "./Language"
+export const keyboardSettings = {
+    languages: [language.ENG, language.RU]
+}
