+++ src/js/Keyboard/constants/Language.js
@@ -1,0 +1,4 @@
+export const language = {
+  RU: 'ru',
+  ENG: 'eng'
+}
